better numerical stability for opaque layers
incoherent calculation now uses the fraction NOT absorbed in a single
pass, rather than the fraction that IS absorbed. This should improve
numerical stability, because the latter and not the former is what
directly enters all the calculations. For a very opaque layer, the
fraction absorbed is very close to 1, so storing it then later
subtracting it from 1 causes loss of significance.